Meeting controler poprawiony

+++ src/main/java/com/company/enroller/persistence/MeetingService.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Component;
 
 import com.company.enroller.model.Meeting;
 
-
 @Component("meetingService")
 public class MeetingService {
 
@@ -24,25 +23,25 @@ public class MeetingService {
 		return query.list();
 	}
 
-	public Meeting findById (int id) {		
+	public Meeting findById(long id) {
 		return (Meeting) connector.getSession().get(Meeting.class, id);
 	}
-	
-	public void create (Meeting meeting) {	
+
+	public void create(Meeting meeting) {
 		Transaction transaction = connector.getSession().beginTransaction();
 		connector.getSession().save(meeting);
 		transaction.commit();
 	}
-	
-	public void delete (Meeting meeting) {	
+
+	public void delete(Meeting meeting) {
 		Transaction transaction = connector.getSession().beginTransaction();
 		connector.getSession().delete(meeting);
 		transaction.commit();
 	}
-	
-	public void update (Meeting meeting) {	
+
+	public void update(Meeting meeting) {
 		Transaction transaction = connector.getSession().beginTransaction();
 		connector.getSession().update(meeting);
 		transaction.commit();
-}
+	}
 }
